refactor: モバイルの死コードと未使用依存を整理 ♻️
