+++ mobile_app/lib/feature/definition_list/application/definition_list_state.dart
