+++ mobile_app/pubspec.yaml
@@ -17,7 +17,6 @@ dependencies:
   app_tracking_transparency: ^2.0.4
   auto_route: ^7.8.3
   cached_network_image: ^3.3.0
-  change_app_package_name: ^1.1.0
   collection: ^1.17.2
   cupertino_icons: ^1.0.2
   device_preview: ^1.1.0
@@ -27,7 +26,6 @@ dependencies:
   flutter_cache_manager: ^3.4.1
   flutter_dotenv: ^5.1.0
   flutter_riverpod: ^2.4.3
-  fluttertoast: ^8.2.3
   freezed_annotation: ^2.4.1
   gap: ^3.0.1
   google_mobile_ads: ^9.0.0
@@ -37,7 +35,6 @@ dependencies:
   intl: ^0.20.2
   keyboard_actions: ^4.2.0
   like_button: ^2.0.5
-  mockito: ^5.4.2
   package_info_plus: ^4.2.0
   pull_down_button: ^0.9.3
   riverpod_annotation: ^2.2.0
@@ -72,6 +69,7 @@ dev_dependencies:
   flutter_lints: ^2.0.0
   flutter_native_splash: ^2.1.3+1
   freezed: ^2.4.1
+  mockito: ^5.4.2
   pedantic_mono: ^1.24.0+1
   riverpod_generator: ^2.3.3
   riverpod_lint: ^2.1.1
